Allow events with individual styles.
For applications with multiple types of events that should be displayed
differently, allow each event to specify its own set of styles for
display.

Calendar now allows an `events` property which is an arry of the form:

`[{date: '2016-07-04', hasEventCircle: {backgroundColor: 'green'}}]`

The existing eventDates property is maintained for backwards
compatibility, events created with eventDates will have the standard
style. Internally events are now passed as objects instead of booleans.